Ignore a directory which can be used to store Docker setup, start, and stop scripts
For convenience, the author of this commit wanted to have a few scripts for
Docker control inside the Coocook directory. Because these scripts partially
contain absolute file paths, they should not be uploaded to origin.

This commit adds an ignore rule for the file path /docker-control, so that one
can freely put some Docker control scripts into a directory with this name.